Make TestOnPRs part of minimal template
The intrusive part of TestOnPRs is the ReusableTest, which is already
added in the minimal. So we might as well benefit from it and have the
improved PR testing.
This removes the question and fix the files accordingly.

Closes #312

Breaking change: Question SimplifiedPRTest is removed. If you chose
recommended or explicitly answered it as true, nothing changes.
If you chose minimum or explicitly answered it as false, you will have a
new file .github/workflows/TestOnPRs.yml and your Test.yml will change.
Pull Requests will run on a smaller matrix now. See the docs for more
information.