Implement debugging tools
Add functions in the debug file to help debug and test the template.
Update the developer documentation to indicate how to use these
functions.
Add a codecov.yml to ignore the debug files.

Closes #316 and #376

Update the developer docs to show how to use the debug tools

Add codecov.yml with ignore for dev files

Small warning