Change the default indentation from 2 to 4
This is the default in the Julia ecosystem, both in the official
guidelines and on various styles.

The Julia files in the template have been changed to use 4 spaces. The
other files in the template are still using 2 spaces. This should be
changed to 4 when pre-commit is run for the first time.

Update the consistency tests that verifies Copier's API to run the
pre-commit before the comparison, to ensure that formatting changes
don't influence the tests.

Closes #403