Support multiple authors
Deprecate AuthorName and AuthorEmail in favour of Authors.
Authors is a comma-separated list of string.
There are no restriction on the format of each string, but
we hope that it will follow the NAME <EMAIL> format.
If that is the case, many places use regex to get the
NAME or EMAIL only and replace the old usage.

A side-effect is that emails are optional now.

Close #118, #356