Don't create main branch when on main and don't run some tests on main
The ReusableTest workflow creates a main branch due a specific test
recently introduced. That specific runs the generation on main and on
HEAD.
Both activites make no sense if we are running on the main branch.

This creates a condition on the workflow to only run if the branch is
not main (a condition on PRs would also work, but we leave the
possibility of running tests on others branches with this approach).
Furthermore, this also adds a condition in the tests such that only on
main the tests are run.

Closes #435