List questions in the documentation
Add a new field to the copier questions: description.
This is not picked up by copier, it is only used in the new docs
page, 30-questions.
Adds a new documentation page 30-questions. This lists the
questions and the new description.
Changes the strategy question to include a link to the questions.

Closes #274