Don't run update tests on releases
The update tests try to compare two versions of the generated
package. Since the release is done form the main branch, this
test can't run properly.
This PR checks if the GitHub CI event is release, and don't
run the test if it is.

Can only be properly tested with a new release.

Related to #454